+++ src/main/java/dev/excal1bur/appliedsmelting/AppliedSmelting.java
@@ -72,6 +72,9 @@ public final class AppliedSmelting {
                     ModBlocks.ME_SMELTER_MK2_ITEM,
                     ModBlocks.ME_SMELTER_MK3_ITEM,
                     ModBlocks.ME_BLAST_FURNACE_ITEM,
+                    ModBlocks.ME_BLAST_FURNACE_MK1_ITEM,
+                    ModBlocks.ME_BLAST_FURNACE_MK2_ITEM,
+                    ModBlocks.ME_BLAST_FURNACE_MK3_ITEM,
                     ModBlocks.ME_SMOKER_ITEM,
                     ModBlocks.ME_CRUCIBLE_ITEM)) {
                 var item = machine.get();